Separate rn-cli config for running from inside GB
