+++ rn-cli.config.js
@@ -1,39 +1,21 @@
 /** @format */
-const path = require( 'path' );
-
-const enm = require( 'node-libs-react-native' );
-enm.fs = __dirname + '/__mocks__/nodejsMock.js';
-enm.net = __dirname + '/__mocks__/nodejsMock.js';
-enm.canvas = __dirname + '/__mocks__/nodejsMock.js';
-
-const { lstatSync, readdirSync } = require( 'fs' );
-const getDirectoryNames = dir =>
-	readdirSync( dir ).filter( name => lstatSync( path.resolve( dir, name ) ).isDirectory() );
-
-const wppackagenames = getDirectoryNames( path.resolve( __dirname, '../packages/' ) );
-
-const mapper = function( accu, v ) {
-	accu[ '@wordpress/' + v ] = path.resolve( __dirname, '../packages/' + v );
-	return accu;
-};
-
-const wppackages = wppackagenames.reduce( mapper, {} );
+const blacklist = require( 'metro' ).createBlacklist;
+const enm = require( './extra-node-modules.config.js' );
 
 module.exports = {
-	extraNodeModules: Object.assign( enm, wppackages ),
-	getProjectRoots() {
-		const roots = [ __dirname, path.resolve( __dirname, '../node_modules' ) ].concat(
-			Object.values( wppackages )
-		);
-		return roots;
+	extraNodeModules: enm,
+	getBlacklistRE: function() {
+		// Blacklist the GB packages we want to consume from NPM (online) directly.
+		// On the other hand, GB packages that are loaded from the source tree directly
+		// are automagically resolved by Metro so, there is no list of them anywhere.
+		return blacklist( [
+			/gutenberg\/packages\/(autop|compose|deprecated|hooks|i18n|is-shallow-equal)\/.*/,
+		] );
 	},
 	getTransformModulePath() {
-		return require.resolve( './sass-transformer-inside-gb.js' );
+		return require.resolve( './sass-transformer.js' );
 	},
 	getSourceExts() {
 		return [ 'js', 'json', 'scss', 'sass' ];
-	},
-	getProvidesModuleNodeModules() {
-		return [ 'react-native-svg', 'react-native' ];
 	},
 };
